Sync iOS and Android samples with the latest win32 PartyNetworkSampleCommon updates (#36)
The intent of this change is to get all of the shared code in PartyNetworkSampleCommon matching between the iOS, Android, and Win32 samples

Though notably I'll call out that I haven't yet re-synced the playfab xplatcppsdk collateral between the 3 samples. Win32 is slightly ahead of the others because of how originally merged in the Win32 sample. I opted not to resync the playfab xplatcppsdk collateral in this change because it would be a LOT of noisey diffs that I think would distract from the more important changes in this PR.

We can sync the playfab xplatcppsdk collateral between the 3 samples in a follow up PR.

While this PR intends to focus on resyncing the shared code in PartyNetworkSampleCommon, a handful of android/ios specific sample files were updated to conform with updated PartyNetworkSampleCommon. There are also a few files which I opportunstically updated from the latest state in our private repo because it seemed like they had gotten out of sync and this public repo's copy was stale.