Update trace instructions to include link to trace file (and other minor improvements) (#43)
The links to our tracing instructions from the MSDN doc pages go directly to the README file. I've gotten feedback once or twice that it wasn't clear where to find the associated tracing script (since the readme just says to use the script but not where to find it which is different depending on the platform).

Also clarifies the command used to run the script which on a fresh windows install might require setting execution policies.